refactor: Better structure the code
_more: Move few things around so it’s clearer what is what.
`mappr.registry` now contains all converter registration code and
`mappr.conversion` is only concerned about the conversion. Move
conversion code outside `TypeConverter`. Makes things
unnecesarily coupled.
