+++ tests/conftest.py
@@ -1,2 +1,2 @@
 # Auto import all fixtures.
-from mappr.testing import *
+from mappr.testing import *     # noqa: F401, F403
