+++ tests/test_mappers.py
@@ -1,7 +1,5 @@
 import dataclasses
 
-import pytest
-
 import mappr
 
 
@@ -17,22 +15,13 @@ class Dst:
     num: int = 20
 
 
-def test_use_default():
-    basic = mappr.TypeConverter(Src, Dst)
-    with_default = mappr.TypeConverter(Src, Dst, mapping=dict(
-        num=mappr.use_default,
-    ))
+def test_auto_generated_converter(scoped_register):
+    mappr.register(Src, Dst)
 
-    src = Src()
-
-    assert basic.convert(src) == Dst(text='hello', num=10)
-    assert with_default.convert(src) == Dst(text='hello', num=20)
+    assert mappr.convert(Dst, Src()) == Dst(text='hello', num=10)
 
 
-def test_cannot_register_converter_twice_for_the_same_types():
-    mappr.register(Src, Dst, mapping=dict(
-        num=mappr.use_default,
-    ))
+def test_use_default(scoped_register):
+    mappr.register(Src, Dst, mapping=dict(num=mappr.use_default))
 
-    with pytest.raises(mappr.ConverterAlreadyExists):
-        mappr.register(Src, Dst)
+    assert mappr.convert(Dst, Src()) == Dst(text='hello', num=20)
